create EvPongTest.cs
