+++ wsnet2-dotnet/WSNet2.Core.Test/EvResponseTest.cs
@@ -5,7 +5,7 @@ using System.Security.Cryptography;
 
 namespace WSNet2.Core.Test
 {
-    public class EvResponsePayloadTests
+    public class EvResponseTests
     {
         MsgPool msgpool;
 
@@ -53,28 +53,6 @@ namespace WSNet2.Core.Test
             Assert.AreEqual(clientDeadline, payload.ClientDeadline);
             Assert.AreEqual(publicProps, payload.PublicProps);
             Assert.AreEqual(privateProps, payload.PrivateProps);
-        }
-
-        [Test]
-        public void TestEvPongPayload()
-        {
-            var payload = new byte[]{
-                (byte)Type.ULong, 1,2,3,4,5,6,7,8,
-                (byte)Type.UInt, 0,0,0,9,
-                (byte)Type.Dict, 2,
-                1, (byte)'a', 0, 9, (byte)Type.ULong, 2,3,4,5,6,7,8,9,
-                2, (byte)'b', (byte)'b', 0, 9, (byte)Type.ULong, 3,4,5,6,7,8,9,10,
-            };
-            var explmts = new Dictionary<string, ulong>(){{"a", 0x0203040506070809}, {"bb", 0x030405060708090a}};
-
-            var reader = WSNet2Serializer.NewReader(payload);
-            var ev = new EvPong(reader);
-            var lmts = new Dictionary<string, ulong>(){{"a", 1}, {"bb", 2}};
-            ev.GetLastMsgTimestamps(lmts);
-
-            Assert.AreEqual(0x0102030405060708, ev.PingTimestamp);
-            Assert.AreEqual(9, ev.WatcherCount);
-            Assert.AreEqual(explmts, lmts);
         }
     }
 }
